% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_immune_tree}
\alias{plot_immune_tree}
\title{Visualize defined immune tree structure}
\usage{
plot_immune_tree()
}
\description{
plot immune tree
}
\examples{
plot_immune_tree()
}
\author{
AXL
}
